requirements.txt
flask>=2.0.0
python-docx>=0.8.11
streamlit>=1.25.0
pandas>=1.5.0
requests>=2.28.0
striprtf>=0.0.26
msal>=1.20.0
requests-oauthlib>=1.3.0
office365-rest-python-client>=2.5.0
